# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CoffeeFlow Backend API",
    "version": "1.0.0",
    "description": "Навчальний REST API. Ціни в копійках UAH; час з часовим поясом. Cookie-сесія та X-CSRFToken. Без JWT і без реєстрації клієнтів."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5000/api/v1",
      "description": "Локальна розробка"
    }
  ],
  "paths": {
...
    "/auth/csrf": {
      "get": {
        "summary": "Отримати CSRF-токен і cookie",
        "description": "Потрібно перед першим записом. Токен діє 1 годину та прив’язаний до cookie.",
        "security": [],
        "responses": {
          "200": {
            "description": "Успішно.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CSRF"
                },
                "example": {
                  "csrf_token": "<токен>"
                }
              }
            }
          },
          "500": {
            "description": "Внутрішня помилка.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "База недоступна; Retry-After: 1.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "parameters": []
      }
    },
...
    "/products": {
      "get": {
        "summary": "Список товарів",
        "description": "Архівні товари виключено; без available повертаються доступні й тимчасово недоступні.",
        "security": [],
        "responses": {
          "200": {
            "description": "Успішно.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductList"
                },
                "example": {
                  "items": [
                    {
                      "id": 1,
                      "category_id": 1,
                      "name": "Еспресо",
                      "description": "Класична чорна кава, 30 мл",
                      "price_cents": 4500,
                      "image_url": "",
                      "is_available": true
                    }
                  ],
                  "total": 1,
                  "limit": 20,
                  "offset": 0
                }
              }
            }
          },
          "422": {
            "description": "Помилка валідації.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Внутрішня помилка.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "База недоступна; Retry-After: 1.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 1000000,
              "default": 0
            }
          },
          {
            "name": "category_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 2147483647
            }
          },
          {
            "name": "available",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          }
        ]
      },
      "post": {
        "summary": "Створити товар",
        "description": "",
        "security": [
          {
            "cookieSession": [],
            "csrfHeader": []
          }
        ],
        "responses": {
          "201": {
            "description": "Успішно.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                },
                "example": {
                  "id": 1,
                  "category_id": 1,
                  "name": "Еспресо",
                  "description": "Класична чорна кава, 30 мл",
                  "price_cents": 4500,
                  "image_url": "",
                  "is_available": true
                }
              }
            },
            "headers": {
              "Location": {
                "description": "URL створеного ресурсу.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "CSRF відсутній / неправильний або пошкоджений JSON.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Потрібен вхід або неправильні облікові дані.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не знайдено або замовлення належить іншому гостю.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Конфлікт даних, недоступний товар або заборонений перехід статусу.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Тіло перевищує 32 КіБ.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "Очікується Content-Type: application/json.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Помилка валідації.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Внутрішня помилка.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "База недоступна; Retry-After: 1.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductInput"
              },
              "example": {
                "category_id": 1,
                "name": "Еспресо",
                "description": "Класична чорна кава, 30 мл",
                "price_cents": 4500,
                "image_url": "",
                "is_available": true
              }
            }
          }
        }
      }
    },
...
    "/orders": {
      "post": {
        "summary": "Створити замовлення",
        "description": "Доступно гостю зі своєю cookie і CSRF-токеном. Ціна з бази; одна транзакція. До 30 запитів за 15 хвилин на IP. Повторний POST створює нове замовлення.",
        "security": [
          {
            "cookieSession": [],
            "csrfHeader": []
          }
        ],
        "responses": {
          "201": {
            "description": "Успішно.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                },
                "example": {
                  "id": 1,
                  "customer_name": "Олена",
                  "phone": "+380501234567",
                  "pickup_at": "2026-09-18T11:00:00+00:00",
                  "comment": "",
                  "total_cents": 9000,
                  "status": "new",
                  "created_at": "2026-09-18T10:00:00+00:00",
                  "items": [
                    {
                      "product_id": 1,
                      "product_name": "Еспресо",
                      "unit_price_cents": 4500,
                      "quantity": 2,
                      "line_total_cents": 9000
                    }
                  ]
                }
              }
            },
            "headers": {
              "Location": {
                "description": "URL створеного ресурсу.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "CSRF відсутній / неправильний або пошкоджений JSON.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Конфлікт даних, недоступний товар або заборонений перехід статусу.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Тіло перевищує 32 КіБ.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "Очікується Content-Type: application/json.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Помилка валідації.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Перевищено ліміт; Retry-After: 900.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Внутрішня помилка.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "База недоступна; Retry-After: 1.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderInput"
              },
              "example": {
                "customer_name": "Олена",
                "phone": "+380501234567",
                "pickup_at": "2026-09-18T14:00:00+03:00",
                "items": [
                  {
                    "product_id": 1,
                    "quantity": 2
                  }
                ]
              }
            }
          }
        }
      },
      "get": {
        "summary": "Список замовлень",
        "description": "",
        "security": [
          {
            "cookieSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Успішно.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderList"
                },
                "example": {
                  "items": [
                    {
                      "id": 1,
                      "customer_name": "Олена",
                      "phone": "+380501234567",
                      "pickup_at": "2026-09-18T11:00:00+00:00",
                      "comment": "",
                      "total_cents": 9000,
                      "status": "new",
                      "created_at": "2026-09-18T10:00:00+00:00",
                      "items": [
                        {
                          "product_id": 1,
                          "product_name": "Еспресо",
                          "unit_price_cents": 4500,
                          "quantity": 2,
                          "line_total_cents": 9000
                        }
                      ]
                    }
                  ],
                  "total": 1,
                  "limit": 20,
                  "offset": 0
                }
              }
            }
          },
          "401": {
            "description": "Потрібен вхід або неправильні облікові дані.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Помилка валідації.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Внутрішня помилка.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "База недоступна; Retry-After: 1.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 1000000,
              "default": 0
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "new",
                "accepted",
                "preparing",
                "ready",
                "completed",
                "cancelled"
              ]
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "fields": {
                "type": "object",
                "additionalProperties": true
              }
            },
            "required": [
              "code",
              "message",
              "fields"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "minimum": 1,
            "maximum": 2147483647
          },
          "category_id": {
            "type": "integer",
            "minimum": 1,
            "maximum": 2147483647
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "default": ""
          },
          "price_cents": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10000000
          },
          "image_url": {
            "type": "string",
            "maxLength": 500,
            "description": "HTTPS URL без username або порожній рядок.",
            "default": ""
          },
          "is_available": {
            "type": "boolean",
            "default": true
          }
        },
        "required": [
          "id",
          "category_id",
          "name",
          "description",
          "price_cents",
          "image_url",
          "is_available"
        ],
        "additionalProperties": false
      },
      "ProductInput": {
        "type": "object",
        "properties": {
          "category_id": {
            "type": "integer",
            "minimum": 1,
            "maximum": 2147483647
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "default": ""
          },
          "price_cents": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10000000
          },
          "image_url": {
            "type": "string",
            "maxLength": 500,
            "description": "HTTPS URL без username або порожній рядок.",
            "default": ""
          },
          "is_available": {
            "type": "boolean",
            "default": true
          }
        },
        "required": [
          "category_id",
          "name",
          "price_cents"
        ],
        "additionalProperties": false
      },
...
      "OrderInput": {
        "type": "object",
        "properties": {
          "customer_name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 80
          },
          "phone": {
            "type": "string",
            "pattern": "^\\+[1-9][0-9]{7,14}$"
          },
          "pickup_at": {
            "type": "string",
            "format": "date-time",
            "description": "Часовий пояс обов’язковий; 15 хвилин–7 днів від поточного UTC часу."
          },
          "comment": {
            "type": "string",
            "maxLength": 500,
            "default": ""
          },
          "items": {
            "type": "array",
            "minItems": 1,
            "maxItems": 30,
            "description": "product_id не повторюються.",
            "items": {
              "type": "object",
              "properties": {
                "product_id": {
                  "type": "integer",
                  "minimum": 1,
                  "maximum": 2147483647
                },
                "quantity": {
                  "type": "integer",
                  "minimum": 1,
                  "maximum": 20
                }
              },
              "required": [
                "product_id",
                "quantity"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "customer_name",
          "phone",
          "pickup_at",
          "items"
        ],
        "additionalProperties": false
      },
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "minimum": 1,
            "maximum": 2147483647
          },
          "customer_name": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "pickup_at": {
            "type": "string",
            "format": "date-time"
          },
          "comment": {
            "type": "string"
          },
          "total_cents": {
            "type": "integer",
            "minimum": 1,
            "maximum": 6000000000
          },
          "status": {
            "type": "string",
            "enum": [
              "new",
              "accepted",
              "preparing",
              "ready",
              "completed",
              "cancelled"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "product_id": {
                  "type": "integer",
                  "minimum": 1,
                  "maximum": 2147483647
                },
                "product_name": {
                  "type": "string"
                },
                "unit_price_cents": {
                  "type": "integer",
                  "minimum": 1,
                  "maximum": 10000000
                },
                "quantity": {
                  "type": "integer",
                  "minimum": 1,
                  "maximum": 20
                },
                "line_total_cents": {
                  "type": "integer",
                  "minimum": 1,
                  "maximum": 200000000
                }
              },
              "required": [
                "product_id",
                "product_name",
                "unit_price_cents",
                "quantity",
                "line_total_cents"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "id",
          "customer_name",
          "phone",
          "pickup_at",
          "comment",
          "total_cents",
          "status",
          "created_at",
          "items"
        ],
        "additionalProperties": false
      },
...
      "CSRF": {
        "type": "object",
        "properties": {
          "csrf_token": {
            "type": "string"
          }
        },
        "required": [
          "csrf_token"
        ],
        "additionalProperties": false
      },
...
      "ProductList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Product"
            }
          },
          "total": {
            "type": "integer",
            "minimum": 0,
            "maximum": 2147483647
          },
          "limit": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          },
          "offset": {
            "type": "integer",
            "minimum": 0,
            "maximum": 1000000
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "additionalProperties": false
      },
      "OrderList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Order"
            }
          },
          "total": {
            "type": "integer",
            "minimum": 0,
            "maximum": 2147483647
          },
          "limit": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          },
          "offset": {
            "type": "integer",
            "minimum": 0,
            "maximum": 1000000
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "cookieSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "coffeeflow_session"
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRFToken"
      }
    }
  }
}